requirements.txt
Django==1.5.4
GitPython==0.3.2.RC1
MySQL-python==1.2.3
PyYAML==3.10
South==0.8.2
WebHDFS==0.2.0
argparse==1.2.1
async==0.6.1
-e git://github.com/KTHse/django-cas2@e8a18601349f551d4949905fd6221e2041fe49c1#egg=django_cas-master
django-dajax==0.9.2
django-dajaxice==0.5.5
django-datatables-view==1.9
django-object-permissions==1.4.6
gitdb==0.5.4
kerberos==1.1.1
pyhs2==0.5.0
python-dateutil==1.5
requests==2.4.3
sasl==0.1.3
smmap==0.8.2
thrift==0.9.1
uWSGI==2.0.8
wsgiref==0.1.2
